Completed refactoring.
